# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for VPN Sell Bot Backend",
        "title": "VPN Sell Bot API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/api/v1",
    "paths": {
        "/orders": {
            "post": {
                "description": "Creates an order and returns payment link if method is crypto",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Orders"
                ],
                "summary": "Create an order",
                "parameters": [
                    {
                        "description": "Order Details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.CreateOrderRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/plans": {
            "get": {
                "description": "Returns all active plans, optionally filtered by server_type (v2ray or wireguard)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Plans"
                ],
                "summary": "Get active plans",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Server type (v2ray or wireguard)",
                        "name": "type",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Plan"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Admin endpoint to create new plans",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Plans"
                ],
                "summary": "Create a new plan",
                "parameters": [
                    {
                        "description": "Plan Details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Plan"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Plan"
                        }
                    }
                }
            }
        },
        "/users": {
            "post": {
                "description": "Fetches a user by telegram_id, creates one if it doesn't exist",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Fetch or create a user",
                "parameters": [
                    {
                        "description": "User Details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.GetOrCreateUserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    }
                }
            }
        },
        "/users/{telegram_id}/balance": {
            "patch": {
                "description": "Adds or subtracts balance from a user's account",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Update user balance",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "Telegram User ID",
                        "name": "telegram_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Amount to update",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.UpdateUserBalanceRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    }
                }
            }
        },
        "/users/{telegram_id}/orders": {
            "get": {
                "description": "Returns all orders for a given Telegram ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Orders"
                ],
                "summary": "Get all user orders",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "Telegram User ID",
                        "name": "telegram_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Order"
                            }
                        }
                    }
                }
            }
        },
        "/users/{telegram_id}/subscriptions": {
            "get": {
                "description": "Returns all subscriptions for a given Telegram ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Subscriptions"
                ],
                "summary": "Get all user subscriptions",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "Telegram User ID",
                        "name": "telegram_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Subscription"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "controllers.CreateOrderRequest": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "payment_method": {
                    "type": "string"
                },
                "plan_id": {
                    "type": "integer"
                },
                "telegram_id": {
                    "type": "integer"
                }
            }
        },
        "controllers.GetOrCreateUserRequest": {
            "type": "object",
            "properties": {
                "language": {
                    "type": "string"
                },
                "telegram_id": {
                    "type": "integer"
                }
            }
        },
        "controllers.UpdateUserBalanceRequest": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                }
            }
        },
        "gorm.DeletedAt": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if Time is not NULL",
                    "type": "boolean"
                }
            }
        },
        "models.Order": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number",
                    "format": "float64"
                },
                "createdAt": {
                    "type": "string"
                },
                "cryptoTxID": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "id": {
                    "type": "integer"
                },
                "paymentMethod": {
                    "description": "\"card\", \"crypto\"",
                    "type": "string"
                },
                "paymentStatus": {
                    "description": "\"pending\", \"approved\", \"rejected\"",
                    "type": "string"
                },
                "planID": {
                    "type": "integer"
                },
                "proofImageID": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "userID": {
                    "type": "integer"
                }
            }
        },
        "models.Plan": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "dataLimitGB": {
                    "type": "number",
                    "format": "float64"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "durationDays": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "isActive": {
                    "type": "boolean"
                },
                "priceIRR": {
                    "type": "number",
                    "format": "float64"
                },
                "priceUSDT": {
                    "type": "number",
                    "format": "float64"
                },
                "serverType": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "models.Subscription": {
            "type": "object",
            "properties": {
                "configLink": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "expiryDate": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "planID": {
                    "type": "integer"
                },
                "serverID": {
                    "type": "integer"
                },
                "startDate": {
                    "type": "string"
                },
                "status": {
                    "description": "\"active\", \"expired\", \"disabled\"",
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "userID": {
                    "type": "integer"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
        "models.User": {
            "type": "object",
            "properties": {
                "balance": {
                    "type": "number"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "id": {
                    "type": "integer"
                },
                "isAdmin": {
                    "type": "boolean"
                },
                "language": {
                    "type": "string"
                },
                "telegramID": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        }
    }
}